+++ src/i18n/zh.js
@@ -876,6 +876,14 @@ export default {
       'end time': '结束时间',
       'start time': '开始时间',
       'Ticket number': '工单编号',
+      Ticket: '工单',
+      number: '编号',
+      type: '类型',
+      status: '状态',
+      'I created': '我创建的',
+      'Assigned to me': '分配给我的',
+      times: '时间',
+      projects: '项目',
       'Ticket type': '工单类型',
       'Ticket status': '工单状态',
       'Ticket description': '工单描述',
